Change CONTRIBUTING.md to advocate using Docker
This can be consistent across the other API clients, some of which
are harder to run natively. Even though we're used to developing in
Python, it's actually quite hard to setup this repo - even without
Tox - so it's worth advocating Docker in the first instance.

Although the task to prepare the image is currently run automatically,
this will change in future commits, so we may as well put in the words
for it now.