Support custom Python versions for local dev
Previously the tracked ".python-version" file forced any local dev
to use the first version in the file, which conflicts with having a
virtualenv [1]. This renames the file in such a way that:

- It still works as before inside a Docker container and on CI. We
still ensure all Python versions are activated with "pyenv global"
instead of a ".python-version" file, which won't exist on CI.

- A local, non-Docker dev needs to make an explicit choice to use
the listed Python versions. I've made this step part of the existing
instructions to install all the Python versions required for tox.

Allowing a local dev to use their preferred Python environment makes
it easier to run a single test locally by avoiding package conflicts
and supporting reuse of existing Python installations.

[1]: https://github.com/alphagov/notifications-manuals/wiki/Getting-Started#development-environment