Fix missing dependencies for local dev
Previously it was hard to run a single test without tox, since the
test dependencies were only listed as part of tox.ini. This replaces
the inline list with a typical "requirements_for_test.txt" file and
changes the "dependencies" task in the Makefile to use that.

Note that the "." syntax in "requirements_for_test.txt" does the same
thing as the previous "pip insall .".

Note also that running tox with the flake8 environment will now install
all the other test dependencies. While this will take a bit longer for
now, we will look at fixing it in the next commits.