Support running commands without tox in Docker
Although tox will automatically install dependencies before it runs
a command, these aren't available normally. Running "make bootstrap"
in the Dockerfile is consistent with our other repos [1], and means
we can e.g. run a single test inside a container.

While the previous "build-with-docker" command would try to install
dependencies, these would only be available in the container for the
"make build" command, which made the command pointless.

Note that I've added a ".dockerignore" file to prevent secrets in a
"environment.sh" file leaking into the build image. In addition, the
".tox" cache directory is often large and not necessary for the image
build, since tox will populate it automatically.

[1]: https://github.com/alphagov/notifications-template-preview/blob/454b89f46e5f6bde145cd08add25a8fe2a127053/docker/Dockerfile#L61-L63