DRY-up running Docker commands with a script
This matches the approach we take in other repos, and has the added
benefit of being able to easily run ad-hoc commands, without needing
to make new tasks for them in the Makefile e.g. individual tests.

Note that I've also removed a couple of niche Docker tasks:

- The task to publish to PyPi can now easily be run using the script,
and is infrequent enough to use inside a Docker container that we don't
need a dedicated task for it.

- The task to cleanup Docker containers should be unnecessary due to
the use of the '--rm' flag.